requirements.txt
numpy
pillow
msgpack
rdkit
torch
torchvision
torch-geometric
matplotlib
